whisper/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
mysql-connector-python
